Move note value strategy
