+++ controlsurfaces/note.py
@@ -8,28 +8,8 @@ Authors:
 """
 from common.types.eventdata import EventData, isEventStandard
 from common.eventpattern import BasicPattern, fromNibbles
-from . import ControlSurface, IValueStrategy
+from . import ControlSurface, NoteStrategy
 
-class NoteValueStrategy(IValueStrategy):
-    """
-    The strategy to get data values from note events
-    """
-    def getValueFromEvent(self, event: EventData) -> int:
-        assert isEventStandard(event)
-        if 0x80 <= event.status < 0x90:
-            return 0
-        else:
-            return event.data2
-    
-    def getChannelFromEvent(self, event: EventData) -> int:
-        assert isEventStandard(event)
-        return event.status & 0xF
-    
-    def getFloatFromValue(self, value: int) -> float:
-        return value / 127
-    
-    def getValueFromFloat(self, f: float) -> int:
-        return int(f * 127)
 
 class Note(ControlSurface):
     """
@@ -38,7 +18,7 @@ class Note(ControlSurface):
     def __init__(self, note_num: int, channel:int = 0) -> None:
         super().__init__(
             BasicPattern(fromNibbles((8, 9), ...), note_num, ...),
-            NoteValueStrategy(),
+            NoteStrategy(),
             "notes",
             (channel, note_num)
         )
